+++ src/main/java/eu/clarin/sru/fcs/qlparser/lex/Relation.java
@@ -73,6 +73,9 @@ public class Relation extends QueryNode {
             }
         }
         sb.append(')');
+        if (location != null) {
+            sb.append('@').append(location.getStart()).append(':').append(location.getStop());
+        }
         return sb.toString();
     }